# admin/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Review Agent admin",
    "version": "1"
  },
  "paths": {
...
    "/api/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Providers",
        "operationId": "providers_api_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/providers/models": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Models",
        "operationId": "models_api_providers_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModelPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/providers/runtime": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Runtime Status",
        "operationId": "runtime_status_api_providers_runtime_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimePage"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/providers/openai-codex/login": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Start Login",
        "operationId": "start_login_api_providers_openai_codex_login_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSession"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/providers/openai-codex/login/{session_id}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Poll Login",
        "operationId": "poll_login_api_providers_openai_codex_login__session_id__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{22,80}$",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/openai-codex/login/{session_id}/cancel": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Cancel Login",
        "operationId": "cancel_login_api_providers_openai_codex_login__session_id__cancel_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{22,80}$",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cancellation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cancellation": {
        "properties": {
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "cancelled",
          "session_id"
        ],
        "title": "Cancellation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HermesRuntimeStatus": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded",
              "unavailable",
              "retrying",
              "unknown"
            ],
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "active_agents": {
            "type": "integer",
            "title": "Active Agents"
          },
          "busy": {
            "type": "boolean",
            "title": "Busy"
          },
          "drainable": {
            "type": "boolean",
            "title": "Drainable"
          },
          "chat_available": {
            "type": "boolean",
            "title": "Chat Available"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/RuntimeCheck"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version",
          "model",
          "active_agents",
          "busy",
          "drainable",
          "chat_available",
          "checks"
        ],
        "title": "HermesRuntimeStatus"
      },
...
      "LoginSession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Code"
          },
          "verification_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Url"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          },
          "poll_interval": {
            "type": "integer",
            "title": "Poll Interval"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "user_code",
          "verification_url",
          "expires_in",
          "poll_interval"
        ],
        "title": "LoginSession"
      },
...
      "ProviderCapability": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "configured",
          "detail"
        ],
        "title": "ProviderCapability"
      },
      "ProviderModel": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "ProviderModel"
      },
      "ProviderModelPage": {
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/ProviderCapability"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderModel"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "items"
        ],
        "title": "ProviderModelPage"
      },
      "ProviderPage": {
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/ProviderCapability"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderStatus"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "items"
        ],
        "title": "ProviderPage"
      },
      "ProviderStatus": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "flow": {
            "type": "string",
            "title": "Flow"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "name",
          "connected",
          "flow",
          "action",
          "expires_at"
        ],
        "title": "ProviderStatus"
      },
...
      "RuntimeCheck": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "state_db",
              "session_store",
              "config",
              "model",
              "disk",
              "gateway",
              "background_queues"
            ],
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded",
              "unavailable",
              "retrying",
              "unknown"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "RuntimeCheck"
      },
      "RuntimePage": {
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/ProviderCapability"
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HermesRuntimeStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "capability",
          "runtime"
        ],
        "title": "RuntimePage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-review_agent_session"
      }
    }
  }
}